Rename component with elect prefix
