+++ example/index.html
@@ -6,16 +6,16 @@
 
     <script type="module">
       import '../dist/staging-banner.js'
-      import '../dist/social-sharer.js'
+      import '../dist/sharer.js'
       import '../dist/footer.js'
-      import '../dist/navigation-bar.js'
+      import '../dist/navbar.js'
     </script>
   </head>
 
   <body>
-    <staging-banner></staging-banner>
-    <navigation-bar></navigation-bar>
-    <social-sharer></social-sharer>
+    <elect-staging-banner></elect-staging-banner>
+    <elect-navbar></elect-navbar>
+    <elect-sharer></elect-sharer>
     <elect-footer></elect-footer>
   </body>
 </html>
